Many changes. Separate Event listeners. New placeholders. New time logging methods in PLMFile, too.
The two classes CommonListener and VanishStatusChangeEventListener will call the old methods in Messages. Messages is no longer a listener.  CommonListener is selected if VNP is not available. This was needed because bukkit returns an error if the VanishStatusChangeEvent exists but the plugin is not enabled. Modified PLM class as it is needed.
Implemented the new placeholders %logins, %totallogins, %uniqueplayers. They are in PLMToolbox, too. Implemented it in Messages yet. Needed methods in PLMFile were added. The new method setPlayerLogin() will register all needed stuff.